+++ src/repo/mongoose/model/order.ts
@@ -2,6 +2,60 @@ import * as mongoose from 'mongoose';
 
 const safe: any = { j: 1, w: 'majority', wtimeout: 10000 };
 
+const customerSchema = new mongoose.Schema(
+    {},
+    {
+        id: false,
+        _id: false,
+        strict: false
+    }
+);
+
+const sellerSchema = new mongoose.Schema(
+    {},
+    {
+        id: false,
+        _id: false,
+        strict: false
+    }
+);
+
+const acceptedOfferSchema = new mongoose.Schema(
+    {},
+    {
+        id: false,
+        _id: false,
+        strict: false
+    }
+);
+
+const paymentMethodSchema = new mongoose.Schema(
+    {},
+    {
+        id: false,
+        _id: false,
+        strict: false
+    }
+);
+
+const discountSchema = new mongoose.Schema(
+    {},
+    {
+        id: false,
+        _id: false,
+        strict: false
+    }
+);
+
+const orderInquiryKeySchema = new mongoose.Schema(
+    {},
+    {
+        id: false,
+        _id: false,
+        strict: false
+    }
+);
+
 /**
  * 注文スキーマ
  * @ignore
@@ -12,26 +66,28 @@ const schema = new mongoose.Schema(
             type: String,
             required: true
         },
-        seller: mongoose.SchemaTypes.Mixed,
-        customer: mongoose.SchemaTypes.Mixed,
+        seller: sellerSchema,
+        customer: customerSchema,
         confirmationNumber: String,
         orderNumber: String,
         price: Number,
         priceCurrency: String,
-        acceptedOffers: [mongoose.SchemaTypes.Mixed],
-        paymentMethods: [mongoose.SchemaTypes.Mixed],
-        discounts: [mongoose.SchemaTypes.Mixed],
+        acceptedOffers: [acceptedOfferSchema],
+        paymentMethods: [paymentMethodSchema],
+        discounts: [discountSchema],
         url: String,
         orderStatus: String,
         orderDate: Date,
         isGift: Boolean,
-        orderInquiryKey: mongoose.SchemaTypes.Mixed
+        orderInquiryKey: orderInquiryKeySchema
     },
     {
         collection: 'orders',
         id: true,
         read: 'primaryPreferred',
         safe: safe,
+        strict: true,
+        useNestedStrict: true,
         timestamps: {
             createdAt: 'createdAt',
             updatedAt: 'updatedAt'
@@ -41,15 +97,13 @@ const schema = new mongoose.Schema(
     }
 );
 
-// 注文番号からの注文取得に使用
-schema.index(
-    { orderNumber: 1 }
-);
-
 // 注文番号はユニークなはず
 schema.index(
     { orderNumber: 1 },
-    { unique: true }
+    {
+        unique: true,
+        name: 'uniqueOrderNumber'
+    }
 );
 
 // 注文照会に使用